# swagger.yaml
openapi: 3.0.0
info:
  title: MEN REST API
  description: MongoDB ExpressJS NodeJS REST API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api/
    description: Localhost development server
components:
  securitySchemes:
    securitySchemes:
      bearerAuth: # arbitrary name for the security scheme
    type: http
    scheme: bearer
    bearerFormat: JWT
  schemas:
    OrganizationModel:
      type: object
      properties:
        orgName:
          type: string
          example: "Acme Corp"
        createdByID:
          type: string
          example: "MongoDBObjectID"
        ownerID:
          type: string
          example: "MongoDBObjectID"
        orgMembers:
          type: array
          example: '["MongoDBObjectID", "MongoDBObjectID"]'
        projectIDs:
          type: array
          example: '["MongoDBObjectID", "MongoDBObjectID"]'
        inviteArray:
          type: array
          example: '["MongoDBObjectID", "MongoDBObjectID"]'
...
    userModel:
      type: object
      properties:
        username:
          type: string
          example: "MattyxD"
        email:
          type: string
          example: "email of the person creating the account"
        fName:
          type: string
          example: "person's first name"
        lName:
          type: string
          example: "person's last name"
        password:
          type: string
          example: "password for the account"
        color:
          type: string
          example: "a hex value which is autogenerated when the user successfully creates their account"
paths:
  /addNewOrganization:
    post:
      tags:
        - organizations
      summary: Add a new organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationModel"
      responses:
        '200':
          description: Organization created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  org:
                    $ref: "#/components/schemas/OrganizationModel"
        '500':
          description: Something went wrong with adding a new Organization
  /updateOrganization/{orgID}:
    post:
      tags:
        - organizations
      summary: Update organization details
      parameters:
        - in: path
          name: orgID
          required: true
          schema:
            type: string
          description: ID of the organization to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationModel"
      responses:
        '200':
          description: Organization updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  organization:
                    $ref: "#/components/schemas/OrganizationModel"
        '404':
          description: Organization not found
        '400':
          description: Invalid organization ID
        '500':
          description: Failed to update the organization
  /getOrganizationsFromID:
    get:
      tags:
        - organizations
      summary: Get organizations associated with the user ID
      responses:
        '200':
          description: Organizations found
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizations:
                    type: array
                    items:
                      $ref: "#/components/schemas/OrganizationModel"
        '500':
          description: Internal Server Error
  /getSpecificOrg/{orgID}:
    get:
      tags:
        - organizations
      summary: Get specific organization details
      parameters:
        - in: path
          name: orgID
          required: true
          schema:
            type: string
          description: ID of the organization to retrieve
      responses:
        '200':
          description: Available Organization
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  org:
                    $ref: "#/components/schemas/OrganizationModel"
        '500':
          description: Internal Server Error
  /check-user-invites:
    get:
      tags:
        - organizations
      summary: Check invitations for the logged-in user
      responses:
        '200':
          description: Invitations found
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    orgName:
                      type: string
                    owner:
                      $ref: "#/components/schemas/userModel"
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /accept-org-inv/{orgID}:
    get:
      tags:
        - organizations
      summary: Accept organization invitation
      parameters:
        - in: path
          name: orgID
          required: true
          schema:
            type: string
          description: ID of the organization invitation to accept
      responses:
        '200':
          description: Invitation accepted
        '401':
          description: Unauthorized
        '404':
          description: Organization not found
        '400':
          description: Invite not found in the organization
        '500':
          description: Internal Server Error
  /decline-org-inv/{orgID}:
    get:
      tags:
        - organizations
      summary: Decline organization invitation
      parameters:
        - in: path
          name: orgID
          required: true
          schema:
            type: string
          description: ID of the organization invitation to decline
      responses:
        '200':
          description: Invitation declined
        '401':
          description: Unauthorized
        '404':
          description: Organization not found
        '400':
          description: Invite not found in the organization
        '500':
          description: Internal Server Error
  /delete-org/{orgID}:
    get:
      tags:
        - organizations
      summary: Delete organization
      parameters:
        - in: path
          name: orgID
          required: true
          schema:
            type: string
          description: ID of the organization to delete
      responses:
        '200':
          description: Organization deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/OrganizationModel"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Organization not found
        '500':
          description: Internal Server Error
...